refactor(browser): centralize sort dialogs and add grid title alignment
- Centralize FolderSortDialog, VideoSortDialog, and FileSystemSortDialog into BrowserSortDialogs.kt to eliminate duplicate implementations
- Add center/left alignment options for folder and video titles in grid view
