+++ app/src/main/java/app/gyrolet/mpvrx/ui/browser/filesystem/FileSystemBrowserScreen.kt
@@ -95,12 +95,8 @@ import app.gyrolet.mpvrx.ui.browser.components.fastScrollGlyph
  import app.gyrolet.mpvrx.ui.browser.dialogs.FileOperationProgressDialog
  import app.gyrolet.mpvrx.ui.browser.dialogs.FolderPickerDialog
  import app.gyrolet.mpvrx.ui.browser.dialogs.RenameDialog
-import app.gyrolet.mpvrx.ui.browser.dialogs.MultiViewModeSelector
-import app.gyrolet.mpvrx.ui.browser.dialogs.SortDialog
-import app.gyrolet.mpvrx.ui.browser.dialogs.ViewModeOption
-import app.gyrolet.mpvrx.ui.browser.dialogs.ViewModeSelector
- import app.gyrolet.mpvrx.ui.browser.dialogs.VisibilityToggle
  import app.gyrolet.mpvrx.ui.browser.dialogs.VideoCompressorOverlay
+import app.gyrolet.mpvrx.ui.browser.dialogs.FileSystemSortDialog
 import app.gyrolet.mpvrx.ui.browser.selection.rememberSelectionManager
 import app.gyrolet.mpvrx.ui.browser.sheets.PlayLinkSheet
 import app.gyrolet.mpvrx.ui.browser.states.EmptyState
@@ -1689,152 +1685,5 @@ private fun FileSystemSearchContent(
   }
 }
 
-@Composable
-fun FileSystemSortDialog(
-  isOpen: Boolean,
-  onDismiss: () -> Unit,
-  isAtRoot: Boolean = true,
-) {
-  val browserPreferences = koinInject<BrowserPreferences>()
-  val appearancePreferences = koinInject<app.gyrolet.mpvrx.preferences.AppearancePreferences>()
-  val folderViewMode by browserPreferences.folderViewMode.collectAsState()
-  val folderSortType by browserPreferences.folderSortType.collectAsState()
-  val folderSortOrder by browserPreferences.folderSortOrder.collectAsState()
-  val showVideoThumbnails by browserPreferences.showVideoThumbnails.collectAsState()
-  val showTotalVideosChip by browserPreferences.showTotalVideosChip.collectAsState()
-  val showTotalSizeChip by browserPreferences.showTotalSizeChip.collectAsState()
-  val showFolderPath by browserPreferences.showFolderPath.collectAsState()
-  val showSizeChip by browserPreferences.showSizeChip.collectAsState()
-  val showResolutionChip by browserPreferences.showResolutionChip.collectAsState()
-  val showFramerateInResolution by browserPreferences.showFramerateInResolution.collectAsState()
-  val showProgressBar by browserPreferences.showProgressBar.collectAsState()
-  val showSubtitleIndicator by browserPreferences.showSubtitleIndicator.collectAsState()
-  val unlimitedNameLines by appearancePreferences.unlimitedNameLines.collectAsState()
 
-  SortDialog(
-    isOpen = isOpen,
-    onDismiss = onDismiss,
-    title = "Sort & View Options",
-    sortType = folderSortType.displayName,
-    onSortTypeChange = { typeName ->
-      app.gyrolet.mpvrx.preferences.FolderSortType.entries.find { it.displayName == typeName }?.let {
-        browserPreferences.folderSortType.set(it)
-      }
-    },
-    sortOrderAsc = folderSortOrder.isAscending,
-    onSortOrderChange = { isAsc ->
-      browserPreferences.folderSortOrder.set(
-        if (isAsc) app.gyrolet.mpvrx.preferences.SortOrder.Ascending
-        else app.gyrolet.mpvrx.preferences.SortOrder.Descending,
-      )
-    },
-    types = listOf(
-      app.gyrolet.mpvrx.preferences.FolderSortType.Title.displayName,
-      app.gyrolet.mpvrx.preferences.FolderSortType.Date.displayName,
-      app.gyrolet.mpvrx.preferences.FolderSortType.Size.displayName,
-    ),
-    icons = listOf(
-      Icons.Filled.Title,
-      Icons.Filled.CalendarToday,
-      Icons.Filled.SwapVert,
-    ),
-    getLabelForType = { type, _ ->
-      when (type) {
-        app.gyrolet.mpvrx.preferences.FolderSortType.Title.displayName -> Pair("A-Z", "Z-A")
-        app.gyrolet.mpvrx.preferences.FolderSortType.Date.displayName -> Pair("Oldest", "Newest")
-        app.gyrolet.mpvrx.preferences.FolderSortType.Size.displayName -> Pair("Smallest", "Largest")
-        else -> Pair("Asc", "Desc")
-      }
-    },
-    showSortOptions = true,
-    viewModeSelector = MultiViewModeSelector(
-      label = "View Mode",
-      options = listOf(
-        ViewModeOption(
-          label = "Folder",
-          icon = Icons.Filled.ViewModule,
-          isSelected = folderViewMode == app.gyrolet.mpvrx.preferences.FolderViewMode.AlbumView,
-          onClick = { browserPreferences.folderViewMode.set(app.gyrolet.mpvrx.preferences.FolderViewMode.AlbumView) }
-        ),
-        ViewModeOption(
-          label = "Tree",
-          icon = Icons.Filled.AccountTree,
-          isSelected = folderViewMode == app.gyrolet.mpvrx.preferences.FolderViewMode.FileManager,
-          onClick = { browserPreferences.folderViewMode.set(app.gyrolet.mpvrx.preferences.FolderViewMode.FileManager) }
-        ),
-        ViewModeOption(
-          label = "Library",
-          icon = Icons.Filled.VideoLibrary,
-          isSelected = folderViewMode == app.gyrolet.mpvrx.preferences.FolderViewMode.MediaLibrary,
-          onClick = { browserPreferences.folderViewMode.set(app.gyrolet.mpvrx.preferences.FolderViewMode.MediaLibrary) }
-        ),
-      )
-    ),
-    layoutModeSelector = ViewModeSelector(
-      label = "Layout",
-      firstOptionLabel = "List",
-      secondOptionLabel = "Grid",
-      firstOptionIcon = Icons.Filled.ViewList,
-      secondOptionIcon = Icons.Filled.GridView,
-      isFirstOptionSelected = true, // Always list mode
-      onViewModeChange = { /* Disabled - do nothing */ },
-    ),
-    folderGridColumnSelector = null,
-    videoGridColumnSelector = null,
-    enableViewModeOptions = isAtRoot,
-    enableLayoutModeOptions = false, // Disabled/grayed out
-    visibilityToggles = listOf(
-      VisibilityToggle(
-        label = "Video Thumbnails",
-        checked = showVideoThumbnails,
-        onCheckedChange = { browserPreferences.showVideoThumbnails.set(it) },
-      ),
-      VisibilityToggle(
-        label = "Full Name",
-        checked = unlimitedNameLines,
-        onCheckedChange = { appearancePreferences.unlimitedNameLines.set(it) },
-      ),
-      VisibilityToggle(
-        label = "Path",
-        checked = showFolderPath,
-        onCheckedChange = { browserPreferences.showFolderPath.set(it) },
-      ),
-      VisibilityToggle(
-        label = "Total Videos",
-        checked = showTotalVideosChip,
-        onCheckedChange = { browserPreferences.showTotalVideosChip.set(it) },
-      ),
-      VisibilityToggle(
-        label = "Folder Size",
-        checked = showTotalSizeChip,
-        onCheckedChange = { browserPreferences.showTotalSizeChip.set(it) },
-      ),
-      VisibilityToggle(
-        label = "Size",
-        checked = showSizeChip,
-        onCheckedChange = { browserPreferences.showSizeChip.set(it) },
-      ),
-      VisibilityToggle(
-        label = "Resolution",
-        checked = showResolutionChip,
-        onCheckedChange = { browserPreferences.showResolutionChip.set(it) },
-      ),
-      VisibilityToggle(
-        label = "Framerate",
-        checked = showFramerateInResolution,
-        onCheckedChange = { browserPreferences.showFramerateInResolution.set(it) },
-      ),
-      VisibilityToggle(
-        label = "Subtitle",
-        checked = showSubtitleIndicator,
-        onCheckedChange = { browserPreferences.showSubtitleIndicator.set(it) },
-      ),
-      VisibilityToggle(
-        label = "Progress Bar",
-        checked = showProgressBar,
-        onCheckedChange = { browserPreferences.showProgressBar.set(it) },
-      ),
-    )
-  )
-}
